requirements.txt
python-socketio
eventlet
mysqlclient
sqlparse
more-itertools
